+++ traits_futures/tests/traits_executor_tests.py
@@ -22,7 +22,7 @@ from traits.api import (
     HasStrictTraits,
     Instance,
     List,
-    on_trait_change,
+    observe,
     Property,
     Tuple,
 )
@@ -101,20 +101,23 @@ class ExecutorListener(HasStrictTraits):
     #: Changes to the 'stopped' trait value.
     stopped_changes = List(Tuple(Bool(), Bool()))
 
-    @on_trait_change("executor:state")
-    def _record_state_change(self, obj, name, old_state, new_state):
+    @observe("executor:state")
+    def _record_state_change(self, event):
+        old_state, new_state = event.old, event.new
         if not self.states:
             # On the first state change, record the initial state as well as
             # the new one.
             self.states.append(old_state)
         self.states.append(new_state)
 
-    @on_trait_change("executor:running")
-    def _record_running_change(self, object, name, old, new):
+    @observe("executor:running")
+    def _record_running_change(self, event):
+        old, new = event.old, event.new
         self.running_changes.append((old, new))
 
-    @on_trait_change("executor:stopped")
-    def _record_stopped_change(self, object, name, old, new):
+    @observe("executor:stopped")
+    def _record_stopped_change(self, event):
+        old, new = event.old, event.new
         self.stopped_changes.append((old, new))
 
 
@@ -127,7 +130,7 @@ class FuturesListener(HasStrictTraits):
     futures = List(Instance(CallFuture))
 
     #: True when all futures have completed.
-    all_done = Property(Bool(), depends_on="futures:done")
+    all_done = Property(Bool(), observe="futures:items:done")
 
     def _get_all_done(self):
         return all(future.done for future in self.futures)
@@ -247,8 +250,8 @@ class TraitsExecutorTests:
             )
 
         results = []
-        future.on_trait_change(
-            lambda result: results.append(result), "result_event"
+        future.observe(
+            lambda event: results.append(event.new), "result_event"
         )
 
         self.wait_until_done(future)
@@ -292,7 +295,7 @@ class TraitsExecutorTests:
         # Triples (state, running, stopped).
         states = []
 
-        def record_states():
+        def record_states(event=None):
             states.append(
                 (
                     self.executor.state,
@@ -301,9 +304,9 @@ class TraitsExecutorTests:
                 )
             )
 
-        self.executor.on_trait_change(record_states, "running")
-        self.executor.on_trait_change(record_states, "stopped")
-        self.executor.on_trait_change(record_states, "state")
+        self.executor.observe(record_states, "running")
+        self.executor.observe(record_states, "stopped")
+        self.executor.observe(record_states, "state")
         submit_call(self.executor, int)
 
         # Record states before, during, and after stopping.